refactor(library): own query-embedding in lib/library, not the AI seam
The library query-embedding logic landed in lib/ai/server.ts only because
GRIDA-SEC-003 confines provider-SDK access to that seam — but that's a reason
for the *provider call* to live there, not the library-specific logic (model
id, 1536-d MRL truncation, normalization, query cache). Those are library
domain: an AI-using feature owned by the Library, not the Library living under
the core AI system.

- Seam keeps a generic `embedTextUnbilled(modelId, value)` primitive (the
  provider-touching part).
- New `editor/lib/library/embedding.ts` owns `embedLibraryQuery` (model id +
  toLibraryVector + cache), calling that primitive.
- search action imports from the library module.
- Drop the unused `embedQueryBilled` (YAGNI — re-add a billed primitive in
  the library module when the agent tool needs it).

Seam audit + typecheck + oxlint clean.
